% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_calc.R
\name{custom_entropy}
\alias{custom_entropy}
\title{Calculate a modified entropy formula.}
\usage{
custom_entropy(vec)
}
\arguments{
\item{vec}{\code{Numeric}. Vector of numeric probabilities.}
}
\description{
This function computes entropy bounded between 0 and 1
using a vector of probabilities.
}
