% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_calc.R
\name{cv2sdlog}
\alias{cv2sdlog}
\title{Converting the coefficient of variation in the original scale to the standard deviation
of the log-normal.}
\usage{
cv2sdlog(x, inverse = FALSE)
}
\arguments{
\item{x}{\code{numeric}. The value of interest.}

\item{inverse}{\code{logical}. Indicates whether to invert the transformation (sdlog to cv)}
}
\description{
This function computes the transformation between the coefficient of variation in
the original scale and the standard deviation of the log normal (sdlog)
}
