% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_provenance_string}
\alias{generate_provenance_string}
\title{This function returns a string that can be added to trace the provenance of a table.}
\usage{
generate_provenance_string(manual_path = NULL, ...)
}
\arguments{
\item{manual_path}{This is a manual path. If not, the function will attempt to use the \code{rstudioapi}
function to get the current path of the active file (might not work outside RStudio).}

\item{...}{Not currently used.}
}
\description{
This function returns a string that can be added to trace the provenance of a table.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{It is recommended that this function be used with a manual path, preferably using the 
function \code{here::here}. 
}\if{html}{\out{</div>}}
}
\examples{
    qtils::generate_provenance_string("/this/is/the/path/t1.1.R")
}
