% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_this_script}
\alias{get_this_script}
\title{Retrieving the path of a script using workarounds}
\usage{
get_this_script(...)
}
\arguments{
\item{...}{Not used.}
}
\description{
Get the directory of the current file either through Rscript or through sourcing within RStudio
}
\references{
https://stackoverflow.com/questions/1815606/determine-path-of-the-executing-script/1816487#1816487
}
\keyword{internal}
