% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_precision_sample_size.R
\name{prec_paireddiff}
\alias{prec_paireddiff}
\title{Quick table for sample size estimation for change-from baseline}
\usage{
prec_paireddiff(
  n = NULL,
  prec = NULL,
  sd1,
  sd2 = NULL,
  cor,
  mean = NULL,
  conf_level = 0.95,
  ...
)
}
\arguments{
\item{n}{\code{integer}. The sample size. Set to NULL to estimate}

\item{prec}{\code{numeric}. The precision. Set to NULL to estimate}

\item{sd1}{\code{numeric}. The standard deviation of the first paired measurement}

\item{sd2}{\code{numeric}. The standard deviation of the second paired measurement.
Defaults to NULL, which sets it equal to \code{sd1}}

\item{cor}{\code{numeric}. The correlation between sd1 and sd2}

\item{mean}{\code{numeric}. Hypothetical mean for confidence interval computation.
Set to 0 by default.}

\item{conf_level}{\code{numeric}. Confidence level.}

\item{...}{Other arguments to \code{presize::prec_mean}}
}
\description{
Quick table for sample size estimation for change-from baseline
}
