% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_precision_sample_size.R
\name{presize_log_backtransform}
\alias{presize_log_backtransform}
\title{Return a \code{presize} object with intervals back-transformed}
\usage{
presize_log_backtransform(presize_obj, ...)
}
\arguments{
\item{presize_obj}{\code{presize}. An object generated by presize.}

\item{...}{Not used}
}
\description{
This is a convenience function to get a log back-transformed
presize object to the original scale.
}
