% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_output_longitudinal.R
\name{qplot_obs_long}
\alias{qplot_obs_long}
\title{Plot observed trajectories of longitudinal data}
\usage{
qplot_obs_long(
  df,
  ptype,
  var,
  tvar,
  id,
  col_by = NULL,
  lt_by = NULL,
  facet_row = NULL,
  facet_col = NULL
)
}
\arguments{
\item{df}{The data frame to plot}

\item{ptype}{\code{string}. This can be either 'mean' or 'spaghetti'}

\item{var}{\code{symbol}. Unquoted variable name representing the primary variable}

\item{tvar}{\code{symbol}. Unquoted variable name representing the primary time variable.}

\item{id}{\code{symbol}. Unquoted variable name representing the identifier column.}

\item{col_by}{\code{symbol}. Unquoted variable name representing the colors. Can be NULL.}

\item{lt_by}{\code{symbol}. Unquoted variable name representing the line type. Can be NULL.}

\item{facet_row}{\code{symbol}. Unquoted variable name representing row facets. Can be NULL}

\item{facet_col}{\code{symbol}. Unquoted variable name representing column facets. Can be NULL}
}
\value{
A \code{ggplot2} plot
}
\description{
This function takes a data frame and plots longitudinal observations
over time. We can plot mean trajectories as well as spaghetti plots.
}
\details{
For mean trajectories, a table representing the number of non-NA observations
will be reported along-side the table. The confidence interval will be computed based on the
normal approximation: \eqn{m + 1.96 * SE}. This function does not label the axes and does not
specify a color scheme or themeing. In other words, themes and labels should be added by the
analyst after.
}
\examples{
data(rand_adlb)
df <- rand_adlb |> dplyr::filter(PARAMCD == "ALT") |> 
    dplyr::mutate(AVISIT = factor(AVISIT))
qplot_obs_long(df = df, ptype = "mean", var = AVAL, 
    tvar = AVISIT, id = USUBJID, col_by = TRT01A)

    
}
