% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_output_simple_models.R
\name{qtab_glm_nb}
\alias{qtab_glm_nb}
\title{Function to provide quick flextable to estimate differences in response rate from a negative binomial model
with offset}
\usage{
qtab_glm_nb(df, var, arm, offset = NULL, covariates = NULL, ...)
}
\arguments{
\item{df}{\code{tibble}/\code{DataFrame} The data frame of interest}

\item{var}{\code{string} Name of the outcome variable}

\item{arm}{\code{string} Name of the primary treatment variable (or any other factor variable)
that is of primary interest}

\item{offset}{\code{symbol}. The offset variable}

\item{covariates}{\code{vector}. A vector of covariates for adjustment}

\item{...}{Additional arguments passed to \code{lm} function}
}
\description{
Function to provide quick flextable to estimate differences in response rate from a negative binomial model
with offset
}
