% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_output_simple_models.R
\name{qtabl_mglm_coef}
\alias{qtabl_mglm_coef}
\title{Table for generating a table combining multiple glm calls}
\usage{
qtabl_mglm_coef(modlist)
}
\arguments{
\item{modlist}{\code{List}. A list of models for combination}
}
\description{
Table for generating a table combining multiple glm calls
}
