% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_output_simple_models.R
\name{quick_lm_ancova_table}
\alias{quick_lm_ancova_table}
\title{Function to provide a quick table for ANCOVA analyses with linear models}
\usage{
quick_lm_ancova_table(df, var, arm, covariates = NULL, ...)
}
\arguments{
\item{df}{(Tibble/DataFrame) The data frame of interest}

\item{var}{(String) Name of the outcome variable}

\item{arm}{(String) Name of the primary treatment variable (or any other factor variable)
that is of primary interest}

\item{covariates}{(Vector of Strings) A vector of covariates for adjustment}

\item{...}{Additional arguments passed to \code{lm} function}
}
\value{
A \code{flextable} table
}
\description{
Function to provide a quick table for ANCOVA analyses with linear models
}
\details{
This function will fit a linear model of the form var ~ arm + covariates
the table will have LS means per strata of \code{arm}, as well as means and
standard deviations of the primary outcome \code{var} stratified by \code{arm}.
The p-value is the ANOVA p-value associated with \code{arm}, since \code{arm}
might have multiple levels.
}
