% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rand_adlb}
\alias{rand_adlb}
\title{Randomly generated ADLB data from the \code{random.cdisc.data} R package}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8400 rows and 102 columns.
}
\usage{
rand_adlb
}
\description{
Seed used to generate 2105 and using default options, with the base
ADSL data being \code{rand_adsl} from this package.
This is a data frame with 400 observations and 55 columns
}
\keyword{datasets}
