% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rand_adsl}
\alias{rand_adsl}
\title{Randomly generated ADSL data from the \code{random.cdisc.data} R package}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 400 rows and 55 columns.
}
\usage{
rand_adsl
}
\description{
Seed used to generate 2105
This is a data frame with 400 observations and 55 columns
}
\keyword{datasets}
