% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_provenance_string}
\alias{generate_provenance_string}
\title{This function returns a string that can be added to trace the provenance of a table.}
\usage{
generate_provenance_string(manual_path = NULL, rmd = FALSE, ...)
}
\arguments{
\item{manual_path}{This is a manual path. If not, the function will attempt to use the \code{rstudioapi}
function to get the current path of the active file (might not work outside RStudio).}

\item{rmd}{(Logical). Indiciate whether an RMD is being rendered. This will affect how the function retrieves
the current file in non-interactive mode.}

\item{...}{Not currently used.}
}
\description{
This function returns a string that can be added to trace the provenance of a table.
}
\details{
It is recommended that this function be used with a manual path, preferably using the
function \code{here}
}
\examples{
    qtils::generate_provenance_string("/this/is/the/path/t1.1.R")
}
